Rename DPT module names for main-DPT number
